Add the hw_feedback_gac_category_to_service_type fn
The primary inspection categories did not give enough detail to identify the type of service in some case. Esp, social care.

So we needed to look at the gac categories as well.

I still have no idea what gac is and there is no list of them on the web. I had to identify them iteratively.