Update to set an excerpt if blank
Using the same process as new services. Create an excerpt and update the post. The value $inspection_category->primary is only set in the object if the category is the primary category. In which case, it is always set to true, so two checks is a bit redundant but you never know.